+++ ios/Video/Dictionary+Merge.swift
@@ -1,0 +1,7 @@
+extension Dictionary {
+    mutating func merge(dict: [Key: Value]) {
+        for (k, v) in dict {
+            updateValue(v, forKey: k)
+        }
+    }
+}