+++ ios/Video/M3U8Kit/Source/NSArray+m3u8.h
@@ -1,0 +1,49 @@
+//
+//  NSArray+m3u8.h
+//  M3U8Kit
+//
+//  Created by Frank on 2022/7/12.
+//  Copyright © 2022 M3U8Kit. All rights reserved.
+//
+// The MIT License (MIT)
+//
+// Copyright (c) 2015 Sun Jin <jeansunvf@gmail.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface NSArray (m3u8)
+
+/** @return "key=value" transform to dictionary */
+- (NSMutableDictionary*)m3u_attributesFromAssignment;
+
+/**
+ If check by invalid value, value will append to last element with specific mark.
+
+ @param mark attribute will be ignored if it is invalid.
+ @return "key=value" transform to dictionary
+ */
+- (NSMutableDictionary*)m3u_attributesFromAssignmentByMark:(nullable NSString*)mark;
+
+@end
+
+NS_ASSUME_NONNULL_END