+++ ios/Video/WeakVideoRef.swift
@@ -1,0 +1,7 @@
+final class WeakVideoRef {
+    weak var video: RCTVideo?
+
+    init(_ video: RCTVideo) {
+        self.video = video
+    }
+}