+++ src/Video.tsx
@@ -66,6 +66,8 @@ export interface VideoRef {
   save: (options: object) => Promise<VideoSaveData>;
   setVolume: (volume: number) => void;
   getCurrentTime: () => Promise<number>;
+  setPrincipalVideoId: (principalId: number) => void;
+  setPeripheralVideoId: (peripheralId: number) => void;
 }
 
 const Video = forwardRef<VideoRef, ReactVideoProps>(
@@ -78,6 +80,8 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
       poster,
       fullscreen,
       drm,
+      principalVideo,
+      peripheralVideo,
       textTracks,
       selectedVideoTrack,
       selectedAudioTrack,
@@ -118,6 +122,8 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
     const nativeRef = useRef<ComponentRef<VideoComponentType>>(null);
     const [showPoster, setShowPoster] = useState(!!poster);
     const [isFullscreen, setIsFullscreen] = useState(fullscreen);
+    const [getPrincipalVideo, setPrincipalVideo] = useState(principalVideo);
+    const [getPeripheralVideo, setPeripheralVideo] = useState(peripheralVideo);
     const [
       _restoreUserInterfaceForPIPStopCompletionHandler,
       setRestoreUserInterfaceForPIPStopCompletionHandler,
@@ -298,6 +304,14 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
 
     const setVolume = useCallback((volume: number) => {
       return VideoManager.setVolume(volume, getReactTag(nativeRef));
+    }, []);
+
+    const setPrincipalVideoId = useCallback((principalId: number) => {
+      setPrincipalVideo(principalId);
+    }, []);
+
+    const setPeripheralVideoId = useCallback((peripheralId: number) => {
+      setPeripheralVideo(peripheralId);
     }, []);
 
     const onVideoLoadStart = useCallback(
@@ -516,6 +530,8 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
         pause,
         resume,
         getCurrentTime,
+        setPrincipalVideoId,
+        setPeripheralVideoId,
         restoreUserInterfaceForPictureInPictureStopCompleted,
         setVolume,
       }),
@@ -527,6 +543,8 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
         pause,
         resume,
         getCurrentTime,
+        setPrincipalVideoId,
+        setPeripheralVideoId,
         restoreUserInterfaceForPictureInPictureStopCompleted,
         setVolume,
       ],
@@ -539,6 +557,8 @@ const Video = forwardRef<VideoRef, ReactVideoProps>(
           {...rest}
           src={src}
           drm={_drm}
+          principalVideo={getPrincipalVideo}
+          peripheralVideo={getPeripheralVideo}
           style={StyleSheet.absoluteFill}
           resizeMode={resizeMode}
           fullscreen={isFullscreen}