+++ src/types/video.ts
@@ -269,6 +269,8 @@ export interface ReactVideoProps extends ReactVideoEvents, ViewProps {
   volume?: number;
   localSourceEncryptionKeyScheme?: string;
   cookiePolicy: EnumValues<CookiePolicy>;
+  principalVideo: number;
+  peripheralVideo: number;
   debug?: DebugConfig;
   allowsExternalPlayback?: boolean; // iOS
   controlsStyles?: ControlsStyles; // Android